requirements.txt
anthropic>=0.18.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.25.0
apscheduler>=3.10.0
